restructure and test setup
