+++ package.json
@@ -35,6 +35,7 @@
     "isparta": "^4.0.0",
     "istanbul": "^0.4.5",
     "mocha": "^3.2.0",
+    "redux-devtools-extension": "^1.0.0",
     "webpack-dev-server": "^1.16.2"
   }
 }
